@@ -1,14 +1,19 @@
-import { NgModule } from '@angular/core';
-import { PhotoComponent } from './photo.component';
-import { CommonModule } from '@angular/common';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule } from "@angular/common/http";
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { PhotoComponent } from "./photo.component";
 
 @NgModule({
-    declarations: [PhotoComponent],
-    imports: [
-        CommonModule,
-        HttpClientModule
-    ],
-    exports: [ PhotoComponent ]
+  declarations:[
+    PhotoComponent
+  ],
+  imports:[
+    CommonModule,
+    HttpClientModule
+  ],
+  exports:[
+    PhotoComponent
+  ]
 })
-export class PhotoModule { }
+
+export class PhotoModule{}
