@@ -1,34 +1,25 @@
-import { NgModule } from '@angular/core';
-import { Routes, RouterModule } from '@angular/router';
-
-import { PhotoListComponent } from './photos/photo-list/photo-list.component';
-import { PhotoFormComponent } from './photos/photo-form/photo-form.component';
-import { NotFoundComponent } from './errors/not-found/not-found.component';
-import { PhotoListResolver } from './photos/photo-list/photo-list.resolver';
+import { NgModule } from "@angular/core";
+import { RouterModule, Routes } from "@angular/router";
+import { NotFoundComponent } from "./errors/not-found/not-found.component";
+import { PhotoFormComponent } from "./photos/photo-form/photo-form.component";
+import { PhotoListComponent } from "./photos/photo-list/photo-list.component";
+import { PhotoListResolver } from "./photos/photo-list/photo-list.resolver";
 
 const routes: Routes = [
-    { 
-        path: 'user/:userName', 
-        component: PhotoListComponent,
-        resolve: {
-            photos: PhotoListResolver
-        }
-    },
-    { 
-        path: 'p/add', 
-        component: PhotoFormComponent 
-    },
-    { 
-        path: '**', 
-        component: NotFoundComponent 
-    }  
+  {
+    path: 'user/:userName',
+    component: PhotoListComponent,
+    resolve:{
+      photos: PhotoListResolver
+    }
+  },
+  { path: 'p/add', component: PhotoFormComponent },
+  { path: '**', component: NotFoundComponent }
 ];
 
 @NgModule({
-    imports: [ 
-        RouterModule.forRoot(routes) 
-    ],
-    exports: [ RouterModule ]
+  imports: [RouterModule.forRoot(routes)],
+  exports: [RouterModule]
 })
-export class AppRoutingModule { }
 
+export class AppRoutingModule { }
