@@ -2,10 +2,12 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NotFoundComponent } from './not-found/not-found.component';
 
+
+
 @NgModule({
+  declarations: [NotFoundComponent],
   imports: [
     CommonModule
-  ],
-  declarations: [NotFoundComponent]
+  ]
 })
 export class ErrorsModule { }
