@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { PhotoFormComponent } from './photo-form.component';
 
@@ -6,12 +6,12 @@ describe('PhotoFormComponent', () => {
   let component: PhotoFormComponent;
   let fixture: ComponentFixture<PhotoFormComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
       declarations: [ PhotoFormComponent ]
     })
     .compileComponents();
-  }));
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(PhotoFormComponent);
