@@ -1,9 +1,12 @@
-import { NgModule } from '@angular/core';
-import { PhotoFormComponent } from './photo-form.component';
-import { CommonModule } from '@angular/common';
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { PhotoFormComponent } from "./photo-form.component";
 
 @NgModule({
-    declarations: [PhotoFormComponent],
-    imports: [ CommonModule ]
+  declarations:[
+    PhotoFormComponent
+  ],
+  imports:[CommonModule]
 })
-export class PhotoFormModule { }
+
+export class PhotoFormModule{}
