@@ -1,22 +1,15 @@
-import { Pipe, PipeTransform } from '@angular/core';
+import { Pipe, PipeTransform } from "@angular/core";
+import { PhotoComponent } from "../photo/photo.component";
 
-import { Photo } from '../photo/photo';
-
-@Pipe({ name: 'filterByDescription'})
+@Pipe({ name: 'filterByDescription' })
 export class FilterByDescription implements PipeTransform {
+  transform(photos: PhotoComponent[], descriptionQuery: string) {
+    descriptionQuery = descriptionQuery.trim().toLowerCase();
 
-    transform(photos: Photo[], descriptionQuery: string) {
-        descriptionQuery = descriptionQuery
-            .trim()
-            .toLowerCase();
-
-        if(descriptionQuery) {
-            return photos.filter(photo => 
-                photo.description.toLowerCase().includes(descriptionQuery)
-            );
-        } else {
-            return photos;
-        }
+    if (descriptionQuery) {
+      return photos.filter(photo => photo.description.toLowerCase().includes(descriptionQuery))
+    } else {
+      return photos;
     }
-
+  }
 }
