@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { LoadButtonComponent } from './load-button.component';
 
@@ -6,12 +6,12 @@ describe('LoadButtonComponent', () => {
   let component: LoadButtonComponent;
   let fixture: ComponentFixture<LoadButtonComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
       declarations: [ LoadButtonComponent ]
     })
     .compileComponents();
-  }));
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(LoadButtonComponent);
