@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { PhotoListComponent } from './photo-list.component';
 
@@ -6,12 +6,12 @@ describe('PhotoListComponent', () => {
   let component: PhotoListComponent;
   let fixture: ComponentFixture<PhotoListComponent>;
 
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
       declarations: [ PhotoListComponent ]
     })
     .compileComponents();
-  }));
+  });
 
   beforeEach(() => {
     fixture = TestBed.createComponent(PhotoListComponent);
