@@ -1,28 +1,28 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-import { PhotoListComponent } from './photo-list.component';
-import { PhotosComponent } from './photos/photos.component';
-import { LoadButtonComponent } from './load-button/load-button.component';
-import { FilterByDescription } from './filter-by-description.pipe';
-import { PhotoModule } from '../photo/photo.module';
-import { CardModule } from '../../shared/components/card/card.module';
-import { SearchComponent } from './search/search.component';
-import { DarkenOnHoverModule } from '../../shared/directives/darken-on-hover/darken-on-hover.module';
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { CardModule } from "src/app/shared/components/card/card.module";
+import { DarkOnHoverDirective } from "src/app/shared/directives/darken-on-hover/darken-on-hover.directive";
+import { PhotoModule } from "../photo/photo.module";
+import { FilterByDescription } from "./filter-by-description.pipe";
+import { LoadButtonComponent } from "./load-button/load-button.component";
+import { PhotoListComponent } from "./photo-list.component";
+import { PhotosComponent } from "./photos/photos.component";
+import { SearchComponent } from "./search/search.component";
 
 @NgModule({
-    declarations: [
-        PhotoListComponent,
-        PhotosComponent,
-        LoadButtonComponent,
-        FilterByDescription,
-        SearchComponent
-    ],
-    imports: [ 
-        CommonModule,
-        PhotoModule,
-        CardModule, 
-        DarkenOnHoverModule
-    ]
+  declarations:[
+    PhotoListComponent,
+    PhotosComponent,
+    LoadButtonComponent,
+    FilterByDescription,
+    SearchComponent,
+    DarkOnHoverDirective
+  ],
+  imports:[
+    CommonModule,
+    PhotoModule,
+    CardModule
+  ]
 })
-export class PhotoListModule {}
+
+export class PhotoListModule{}
