@@ -1,30 +1,29 @@
-import { Component, Input, OnChanges, SimpleChanges } from '@angular/core';
-
-import { Photo } from '../../photo/photo';
-
+import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
+import { PhotoComponent } from "../../photo/photo.component"
 @Component({
   selector: 'ap-photos',
   templateUrl: './photos.component.html',
   styleUrls: ['./photos.component.css']
 })
 export class PhotosComponent implements OnChanges {
-  
-  @Input() photos: Photo[] = [];
-  rows: any[] = [];
-  
+
+  @Input() photos: PhotoComponent[] = [];
+  rows = [];
+
   constructor() { }
-  
+
   ngOnChanges(changes: SimpleChanges) {
-    if(changes.photos) 
-      this.rows = this.groupColumns(this.photos);
+    if(changes.photos)
+    this.rows = this.groupColumns(this.photos);
   }
 
-  groupColumns(photos: Photo[]) {
+  groupColumns(photos){
     const newRows = [];
 
-    for(let index = 0; index < photos.length; index+=3) {
-      newRows.push(photos.slice(index, index + 3));
-    }                            
+    for(let i = 0; i<photos.length; i++){
+      newRows.push(photos.slice(i,i+3))
+    };
+
     return newRows;
   }
 }
