@@ -1,12 +1,11 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input } from "@angular/core";
 
 @Component({
-    selector: 'ap-photo',
-    templateUrl: 'photo.component.html'
+  selector: 'ap-photo',
+  templateUrl: './photo.component.html'
 })
+
 export class PhotoComponent {
-    
-    @Input() description='';
-    
-    @Input() url='';
+  @Input() url = '';
+  @Input() description = '';
 }
