@@ -1,10 +1,10 @@
-export interface Photo {
-    id:number;
-    postDate:Date;
-    url:string;
-    description:string;
-    allowComments:boolean;
-    likes:number;
-    comments:number;
-    userId:number;     
+export interface Photo{
+  id: number,
+  postDate: Date,
+  url: string,
+  description: string,
+  allowComments: boolean,
+  likes: number,
+  comments: number,
+  userId: number
 }
