@@ -1,16 +1,13 @@
-import { NgModule } from '@angular/core';
-import { PhotoModule } from './photo/photo.module';
-import { PhotoFormModule } from './photo-form/photo-form.module';
-import { PhotoListModule } from './photo-list/photo-list.module';
-import { CardModule } from '../shared/components/card/card.module';
-import { DarkenOnHoverModule } from '../shared/directives/darken-on-hover/darken-on-hover.module';
+import { NgModule } from "@angular/core";
+import { PhotoListModule } from "./photo-list/photo-list.module";
+import { PhotoModule } from "./photo/photo.module";
+import { PhotoFormModule } from "./photo-form/photo-form.module";
 
 @NgModule({
-    imports: [ 
-        PhotoModule,
-        PhotoFormModule,
-        PhotoListModule,
-        DarkenOnHoverModule
-    ]
+  imports: [
+    PhotoModule,
+    PhotoFormModule,
+    PhotoListModule
+  ]
 })
-export class PhotosModule {}
+export class PhotosModule { }
