@@ -1,9 +1,9 @@
-import { Component, Input } from '@angular/core';
+import { Component, Input } from "@angular/core";
 
 @Component({
-    selector: 'ap-card',
-    templateUrl: './card.component.html'
+  selector: 'ap-card',
+  templateUrl: './card.component.html'
 })
-export class CardComponent { 
-    @Input() title: string = '';
+export class CardComponent{
+  @Input() title:string = '';
 }
