@@ -1,10 +1,10 @@
-import { NgModule } from '@angular/core';
-import { CardComponent } from './card.component';
-import { CommonModule } from '@angular/common';
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { CardComponent } from "./card.component";
 
 @NgModule({
-    declarations: [CardComponent],
-    exports: [CardComponent],
-    imports: [CommonModule]
+  declarations: [CardComponent],
+  exports: [CardComponent],
+  imports: [CommonModule]
 })
 export class CardModule { }
