@@ -1,24 +1,23 @@
-import { Directive, ElementRef, HostListener, Renderer, Input } from '@angular/core';
+import { Directive, ElementRef, HostListener, Renderer2 } from "@angular/core";
 
 @Directive({
-    selector: '[apDarkenOnHover]'
+  selector: '[apDarkOnHover]'
 })
-export class DarkenOnHoverDirective { 
+export class DarkOnHoverDirective {
+  constructor(
+    private el: ElementRef,
+    private render: Renderer2
+    ) {
 
-    @Input() brightness = '70%';
+  }
 
-    constructor(
-        private el: ElementRef,
-        private render: Renderer
-    ) {}
+  @HostListener('mouseover')
+  darkenOn(){
+    this.render.setStyle(this.el.nativeElement,'filter','brightness(70%)')
+  }
 
-    @HostListener('mouseover')
-    darkenOn() {
-        this.render.setElementStyle(this.el.nativeElement, 'filter', `brightness(${this.brightness})`);
-    }
-
-    @HostListener('mouseleave')
-    darkenOff() {
-        this.render.setElementStyle(this.el.nativeElement, 'filter', 'brightness(100%)');
-    }
+  @HostListener('mouseleave')
+  darkenOff(){
+    this.render.setStyle(this.el.nativeElement,'filter','brightness(100%)')
+  }
 }
